fix: type 수정
